# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For OrderSpack PHP Test",
        "description": "This is a demonstration of OrderSpack Php test.",
        "contact": {
            "name": "Developer",
            "url": "https://boosurostephen.com",
            "email": "boosurostephen@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "PHP TEST API Server"
        }
    ],
    "paths": {
...
        "/api/v1/productvariants": {
            "get": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Fetch All Product Variants",
                "description": "Get list of product variants",
                "operationId": "getAllProductVariants",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Store a product variant",
                "description": "Store Product Variant",
                "operationId": "storeProductVariant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "variant_name": {
                                        "description": "Product Variant",
                                        "type": "string",
                                        "example": "XL"
                                    },
                                    "variant_group_id": {
                                        "description": "Variant Group Id",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "product variant created successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/productvariants/{id}": {
            "get": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Fetch a Product Variant",
                "description": "Fetch a product variant",
                "operationId": "getProductVariant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Variant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "product variant retrieved successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Update a product variant",
                "description": "Update Product Variant",
                "operationId": "updateProductVariant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Variant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "variant_name",
                        "in": "query",
                        "description": "Product Variant Name",
                        "schema": {
                            "type": "string"
                        },
                        "example": "XL"
                    },
                    {
                        "name": "variant_group_id",
                        "in": "query",
                        "description": "Variant Group Id",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "product variant updated successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product Variants"
                ],
                "summary": "Delete Product Variant",
                "description": "Delete a product variant",
                "operationId": "deleteProductVariant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Variant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product Variant deleted successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}